runtime: isolate wasm cleanup lifecycle from bare metal
